% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grams.R
\name{text_grams}
\alias{text_grams}
\title{Created on 09 February, 2019
auuthor: Harjyot Kaur
Implementation of text_grams function in the RySyntext package.
This function returns a DataFrame with k top ngrams.
ngrams is a combination of n words occuring together with
highest frequency in the text. The function will return multiple
values in cases of frequency conflict.
Takes in a string and returns a data.frame
Number of rows are dependent on the input n of the user
Size of the list is dependent on the input k of the user}
\usage{
text_grams(text)
}
\arguments{
\item{text}{string}

\item{k}{top ngrams}

\item{n}{n combination of words}

\item{stopwords_remove}{Boolean}

\item{lemmatize}{Boolean}

\item{remove_punctuation}{Boolean}

\item{remove_numbers}{Boolean}

\item{case_sensitive}{Boolean}

\item{gibberish_remove}{Boolean}
}
\value{
data.frame
}
\description{
Created on 09 February, 2019
auuthor: Harjyot Kaur
Implementation of text_grams function in the RySyntext package.
This function returns a DataFrame with k top ngrams.
ngrams is a combination of n words occuring together with
highest frequency in the text. The function will return multiple
values in cases of frequency conflict.
Takes in a string and returns a data.frame
Number of rows are dependent on the input n of the user
Size of the list is dependent on the input k of the user
}
\examples{
text <-  "Today is a sunny day. We should go to a beach on this sunny day"

grams<- text_grams(text)
}
